+++ purchases/src/defaults/kotlin/com/revenuecat/purchases/coroutinesExtensions.kt
@@ -4,6 +4,7 @@ import com.revenuecat.purchases.CacheFetchPolicy.CACHED_OR_FETCHED
 import com.revenuecat.purchases.customercenter.CustomerCenterConfigData
 import com.revenuecat.purchases.data.LogInResult
 import com.revenuecat.purchases.interfaces.GetCustomerCenterConfigCallback
+import com.revenuecat.purchases.virtualcurrencies.VirtualCurrencies
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
@@ -188,5 +189,26 @@ suspend fun Purchases.awaitCustomerCenterConfigData(): CustomerCenterConfigData 
                 continuation.resumeWithException(PurchasesException(error))
             }
         })
+    }
+}
+
+/**
+ * Fetches the virtual currencies for the current subscriber.
+ *
+ * Coroutine friendly version of [Purchases.getVirtualCurrencies].
+ *
+ * @throws [PurchasesException] with a [PurchasesError] if an error occurred while fetching
+ * the virtual currencies.
+ *
+ * @return The [VirtualCurrencies] with the subscriber's virtual currencies.
+ */
+@JvmSynthetic
+@Throws(PurchasesException::class)
+suspend fun Purchases.awaitGetVirtualCurrencies(): VirtualCurrencies {
+    return suspendCoroutine { continuation ->
+        getVirtualCurrenciesWith(
+            onSuccess = { continuation.resume(it) },
+            onError = { continuation.resumeWithException(PurchasesException(it)) }
+        )
     }
 }
